# specs/api/canvas-api.openapi.yml
openapi: 3.0.3
info:
  title: Canvas Learning System API
  description: |
    Canvas学习系统 - Obsidian Canvas学习系统API

    本API覆盖Canvas Learning System的四层架构：
    - Layer 1: CanvasJSONOperator - 基础JSON CRUD操作
    - Layer 2: CanvasBusinessLogic - 业务逻辑层
    - Layer 3: CanvasOrchestrator - 编排层，调度Sub-agents
    - Layer 4: KnowledgeGraphLayer - 知识图谱层

    **核心规范**:
    - 所有Canvas操作必须保持文件完整性
    - 颜色代码: "1"(红), "2"(橙), "3"(黄), "5"(绿), "6"(紫)
    - UTF-8编码，支持中文

    **版本**: 2.0.0
    **更新日期**: 2025-11-19
  version: 2.0.0
  x-source-verification:
    verified_at: "2025-11-25T18:25:00Z"
    format_source:
      type: context7
      library_id: "/oai/openapi-specification"
      topic: "info object, paths, components, specification extensions"
    business_source:
      prd_version: "v1.1.9"
      epic: "Epic 11"
      story_refs: ["11.1", "11.2", "11.3", "11.4", "11.5"]
  contact:
    name: Canvas Learning System Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v2
    description: Local development server
...
paths:
...
  /canvas/{canvas_path}:
    get:
      summary: 读取Canvas文件
      operationId: readCanvas
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          description: Canvas文件路径
          schema:
            type: string
      responses:
        '200':
          description: 成功返回Canvas数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasData'
    put:
      summary: 写入Canvas文件
      operationId: writeCanvas
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanvasData'
      responses:
        '200':
          description: 成功写入Canvas文件
  /canvas/{canvas_path}/nodes:
    post:
      summary: 添加节点到Canvas
      operationId: addNode
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeInput'
      responses:
        '201':
          description: 成功添加节点
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasNode'
  /canvas/{canvas_path}/nodes/{node_id}:
    get:
      summary: 获取单个节点
      operationId: getNode
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功返回节点
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasNode'
    put:
      summary: 更新节点
      operationId: updateNode
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeInput'
      responses:
        '200':
          description: 成功更新节点
    delete:
      summary: 删除节点
      operationId: deleteNode
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 成功删除节点

  # Layer 2: CanvasBusinessLogic
...
  /canvas/{canvas_path}/edges:
    post:
      summary: 创建边（连接）
      operationId: createEdge
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeInput'
      responses:
        '201':
          description: 成功创建边
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasEdge'
  /canvas/{canvas_path}/edges/{edge_id}:
    delete:
      summary: 删除边
      operationId: deleteEdge
      tags: [Layer1-JSON]
      parameters:
        - name: canvas_path
          in: path
          required: true
          schema:
            type: string
        - name: edge_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 成功删除边

  # Story 36.4: Canvas Edge Bulk Sync to Neo4j
  /canvas/{canvas_path}/sync-edges:
    post:
      summary: 同步Canvas所有edges到Neo4j
      operationId: syncCanvasEdges
      tags: [Layer4-Knowledge]
      description: |
        同步Canvas所有edges到Neo4j知识图谱。

        **特性**:
        - 幂等操作: 使用MERGE语法，重复调用不创建重复关系
        - 并发处理: 最多12个edge同时同步
        - 部分失败: 单个edge失败不阻塞整体批处理
        - 性能目标: 100 edges < 5秒

        **Story 36.4**: Canvas打开时全量Edge同步
        [Source: docs/stories/36.4.story.md]
      parameters:
        - name: canvas_path
          in: path
          required: true
          description: Canvas文件路径（不含.canvas扩展名）
          schema:
            type: string
            example: "离散数学"
      responses:
        '200':
          description: 同步完成，返回摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncEdgesSummary'
        '404':
          description: Canvas文件不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Canvas not found: 离散数学"

  # 回滚操作
...
components:
  schemas:
    CanvasData:
      type: object
      description: Canvas文件数据结构
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/CanvasNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/CanvasEdge'
      required: [nodes, edges]
    CanvasNode:
      type: object
      description: Canvas节点
      properties:
        id:
          type: string
          description: 节点唯一标识
        type:
          type: string
          enum: [text, file, group, link]
          description: 节点类型
        text:
          type: string
          description: 文本内容（type=text时）
        file:
          type: string
          description: 文件路径（type=file时）
        url:
          type: string
          description: URL（type=link时）
        x:
          type: integer
          description: X坐标
        y:
          type: integer
          description: Y坐标
        width:
          type: integer
          description: 宽度
        height:
          type: integer
          description: 高度
        color:
          type: string
          enum: ["1", "2", "3", "4", "5", "6"]  # Story 12.B.4
          description: "颜色代码: 1=红, 2=橙, 3=黄, 5=绿, 6=紫"
      required: [id, type, x, y]
    CanvasEdge:
      type: object
      description: Canvas边（连接）
      properties:
        id:
          type: string
          description: 边唯一标识
        fromNode:
          type: string
          description: 起始节点ID
        toNode:
          type: string
          description: 目标节点ID
        fromSide:
          type: string
          enum: [top, bottom, left, right]
        toSide:
          type: string
          enum: [top, bottom, left, right]
      required: [id, fromNode, toNode]
    NodeInput:
      type: object
      description: 节点输入数据
      properties:
        type:
          type: string
          enum: [text, file, group, link]
        text:
          type: string
        file:
          type: string
        url:
          type: string
        x:
          type: integer
        y:
          type: integer
        width:
          type: integer
        height:
          type: integer
        color:
          type: string
          enum: ["1", "2", "3", "4", "5", "6"]  # Story 12.B.4
      required: [type]
...
    EdgeInput:
      type: object
      description: 边输入数据
      properties:
        fromNode:
          type: string
          description: 起始节点ID
        toNode:
          type: string
          description: 目标节点ID
        fromSide:
          type: string
          enum: [top, bottom, left, right]
          description: 起始节点连接侧
        toSide:
          type: string
          enum: [top, bottom, left, right]
          description: 目标节点连接侧
        color:
          type: string
          description: 边颜色（可选）
        label:
          type: string
          description: 边标签（可选）
      required: [fromNode, toNode]

    # Story 36.4: Edge Bulk Sync Summary
    SyncEdgesSummary:
      type: object
      description: |
        Edge同步结果摘要。

        Story 36.4: Canvas打开时全量Edge同步
        [Source: docs/stories/36.4.story.md#Task-5.2]
      properties:
        canvas_path:
          type: string
          description: Canvas文件路径
          example: "离散数学"
        total_edges:
          type: integer
          description: Canvas中的总边数
          example: 50
        synced_count:
          type: integer
          description: 成功同步到Neo4j的边数
          example: 48
        failed_count:
          type: integer
          description: 同步失败的边数
          example: 2
        skipped_count:
          type: integer
          description: 跳过的边数（如memory client不可用）
          example: 0
        sync_time_ms:
          type: number
          format: float
          description: 同步耗时（毫秒）
          example: 1250.5
      required:
        - canvas_path
        - total_edges
        - synced_count
        - failed_count
        - skipped_count
        - sync_time_ms
      example:
        canvas_path: "离散数学"
        total_edges: 50
        synced_count: 50
        failed_count: 0
        skipped_count: 0
        sync_time_ms: 1250.5
...